New item view (#11)
* imported necessary modules, created NewItemViewViewModel class skeleton

* add title, dueDate, showAlert properties

title, dueDate is name and due date of new Task, and showAlert will determine if alert should be shown. Any changes will be notified to observers

* add canSave() function

checks if the entered title, date for task is not empty, not in past

* add init function

* add save() function

saves entered task by user to the firebase db. code involving model will be uncommented when model code available.

* add vars viewModel, newItemPresented

viewModel will be state object of NewItemViewViewModel, and newItemPresented will be binding of it and the var given to NewItemView

* refactored preview to have newItemPresented parameter

* add 'New Item' text

* add form for title, and due date of task

commented out tl-button code until required code becomes available

* changed typo 'ToDoListApp' to 'ToDoListItem'

* uncommented tl-button code

* uncommented code involving ToDoListItem